Move DownloadData to a separate package.
